+++ golem-worker-executor-base/src/workerctx.rs
@@ -292,7 +292,7 @@ pub trait ExternalOperations<Ctx: WorkerCtx> {
     async fn get_worker_retry_count<T: HasAll<Ctx> + Send + Sync>(
         this: &T,
         worker_id: &WorkerId,
-    ) -> u32;
+    ) -> u64;
 
     /// Gets a best-effort current worker status without activating the worker
     async fn get_assumed_worker_status<T: HasAll<Ctx> + Send + Sync>(
