+++ golem-worker-service/src/service/worker.rs
@@ -217,7 +217,7 @@ pub trait WorkerService {
     async fn complete_promise(
         &self,
         worker_id: &WorkerId,
-        oplog_id: i32,
+        oplog_id: u64,
         data: Vec<u8>,
     ) -> Result<bool, WorkerError>;
 
@@ -766,7 +766,7 @@ impl WorkerService for WorkerServiceDefault {
     async fn complete_promise(
         &self,
         worker_id: &WorkerId,
-        oplog_id: i32,
+        oplog_id: u64,
         data: Vec<u8>,
     ) -> Result<bool, WorkerError> {
         let promise_id = PromiseId {
@@ -1013,7 +1013,7 @@ impl WorkerService for WorkerServiceNoOp {
     async fn complete_promise(
         &self,
         _worker_id: &WorkerId,
-        _oplog_id: i32,
+        _oplog_id: u64,
         _data: Vec<u8>,
     ) -> Result<bool, WorkerError> {
         Ok(true)
